+++ src/aux-server/aux-web/webpack.common.js
@@ -121,7 +121,19 @@ module.exports = {
                 include: [/aux-common/, /aux-vm/],
                 enforce: 'pre',
             },
+
+            // See https://github.com/dchest/tweetnacl-js/wiki/Using-with-Webpack
+            // Gist is that tweetnacl-js has some require() statements that webpack
+            // will parse and may try to include shims automatically.
+            // So here we tell webpack to ignore tweetnacl and import from the global
+            // window.nacl property.
+            {
+                test: /[\\\/]tweetnacl[\\\/]/,
+                loader:
+                    'exports-loader?globalThis.nacl!imports-loader?this=>globalThis,module=>{},require=>false',
+            },
         ],
+        noParse: [/[\\\/]tweetnacl[\\\/]/, /[\\\/]tweetnacl-auth[\\\/]/],
     },
     resolve: {
         extensions: ['.vue', '.js', '.ts', '.css'],