+++ src/crypto/index.ts
@@ -2,3 +2,4 @@ export * from './SignatureAlgorithm';
 export * from './CryptoImpl';
 export * from './utils';
 export * from './HashHelpers';
+export * from './Encryption';