+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # CasualOS Changelog
 
-## V1.1.19
+## V1.2.0
 
-### Date: TBD
+### Date: 7/17/2020
 
 ### Changes:
 
@@ -24,6 +24,18 @@
             -   `key` is the password that should be used for the message authentication code.
             -   `data` is the data to calculate the HMAC of.
             -   Supports strings, numbers, booleans, objects, arrays, and bots.
+        -   `crypto.encrypt(password, data)`
+            -   Encrypts the given data with the given password and returns the result as a promise.
+            -   `password` is the password to use for encrypting the data.
+            -   `data` is the data that should be encrypted.
+        -   `crypto.decrypt(password, data)`
+            -   Decrypts the given data with the given password and returns the result as a promise.
+            -   Only works if the given data is the output of `crypto.encrypt()`.
+            -   `password` is the password that was used to encrypt the data.
+            -   `data` is the data that should be decrypted.
+
+-   :bug: Bug Fixes
+    -   Fixed a race condition where concurrently updating a tag in a script and triggering a dependency update on that same tag could cause the runtime to crash.
 
 ## V1.1.18